+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Change history log to include archive mounting errors
 - Change behavior on missing keyring daemon: Save passwords in memory until close
 - Change to offer to stop browsing archive for every borg task
+- Change file and folder choosers to preselect a folder
 - Add docker to containers and VMs exclusion preset
 - Add per-backup preferences window
 - Add ability to set custom titles for backup configs