+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Change to offer to stop browsing archive for every borg task
 - Change file and folder choosers to preselect a folder
 - Change backup notifications to be dismissed when a new backup starts
+- Change fnmatches to be stored as strings instead of bytes in config
 - Add docker to containers and VMs exclusion preset
 - Add per-backup preferences window
 - Add ability to set custom titles for backup configs